requirements.txt
flask==0.12.2
PyYAML==5.1
requests==2.19.1
